CurrentSegment Double, now runs from Segment Seperator

+++ JHotDraw/src/main/java/org/jhotdraw/geom/SegmentSeperator.java
@@ -10,16 +10,17 @@ import static java.awt.geom.PathIterator.SEG_QUADTO;
  * @author Firefigher
  */
 public class SegmentSeperator implements PathIterator {
+
     float[] floatCoords;
     double[] doubleCoords;
     int numCoords;
     int type;
-    
 
     public SegmentSeperator() {
     }
-    
+
     public void lineToOrQuadto(float[] coords, BezierPath.Node current, int numCoords, int type) {
+            
         if ((current.mask & BezierPath.C1_MASK) == 0) {
                         numCoords = 1;
                         type = SEG_LINETO;
@@ -33,85 +34,83 @@ public class SegmentSeperator implements PathIterator {
                         coords[2] = (float) current.x[0];
                         coords[3] = (float) current.y[0];
                     }
+         
         this.numCoords = numCoords;
         this.type = type;
+         
         this.floatCoords = coords;
     }
-    
+
     public void quadtoOrCubicto(float[] coords, BezierPath.Node current, BezierPath.Node previous, int numCoords, int type) {
         
-        
         if ((current.mask & BezierPath.C1_MASK) == 0) {
-                        numCoords = 2;
-                        type = SEG_QUADTO;
-                        coords[0] = (float) previous.x[2];
-                        coords[1] = (float) previous.y[2];
-                        coords[2] = (float) current.x[0];
-                        coords[3] = (float) current.y[0];
-                    } else {
-                        numCoords = 3;
-                        type = SEG_CUBICTO;
-                        coords[0] = (float) previous.x[2];
-                        coords[1] = (float) previous.y[2];
-                        coords[2] = (float) current.x[1];
-                        coords[3] = (float) current.y[1];
-                        coords[4] = (float) current.x[0];
-                        coords[5] = (float) current.y[0];
-                    }
+            numCoords = 2;
+            type = SEG_QUADTO;
+            coords[0] = (float) previous.x[2];
+            coords[1] = (float) previous.y[2];
+            coords[2] = (float) current.x[0];
+            coords[3] = (float) current.y[0];
+        } else {
+            numCoords = 3;
+            type = SEG_CUBICTO;
+            coords[0] = (float) previous.x[2];
+            coords[1] = (float) previous.y[2];
+            coords[2] = (float) current.x[1];
+            coords[3] = (float) current.y[1];
+            coords[4] = (float) current.x[0];
+            coords[5] = (float) current.y[0];
+        }
         this.numCoords = numCoords;
         this.type = type;
         this.floatCoords = coords;
     }
 
-    
     public void lineToOrQuadto(double[] coords, BezierPath.Node current, int numCoords, int type) {
+
         if ((current.mask & BezierPath.C1_MASK) == 0) {
-                        numCoords = 1;
-                        type = SEG_LINETO;
-                        coords[0] = (float) current.x[0];
-                        coords[1] = (float) current.y[0];
-                    } else {
-                        numCoords = 2;
-                        type = SEG_QUADTO;
-                        coords[0] = (float) current.x[1];
-                        coords[1] = (float) current.y[1];
-                        coords[2] = (float) current.x[0];
-                        coords[3] = (float) current.y[0];
-                    }
+            numCoords = 1;
+            type = SEG_LINETO;
+            coords[0] = current.x[0];
+            coords[1] = current.y[0];
+        } else {
+            numCoords = 2;
+            type = SEG_QUADTO;
+            coords[0] = current.x[1];
+            coords[1] = current.y[1];
+            coords[2] = current.x[0];
+            coords[3] = current.y[0];
+        }
         this.numCoords = numCoords;
         this.type = type;
         this.doubleCoords = coords;
     }
-    
+
     public void quadtoOrCubicto(double[] coords, BezierPath.Node current, BezierPath.Node previous, int numCoords, int type) {
-        
-        
+
         if ((current.mask & BezierPath.C1_MASK) == 0) {
-                        numCoords = 2;
-                        type = SEG_QUADTO;
-                        coords[0] = (float) previous.x[2];
-                        coords[1] = (float) previous.y[2];
-                        coords[2] = (float) current.x[0];
-                        coords[3] = (float) current.y[0];
-                    } else {
-                        numCoords = 3;
-                        type = SEG_CUBICTO;
-                        coords[0] = (float) previous.x[2];
-                        coords[1] = (float) previous.y[2];
-                        coords[2] = (float) current.x[1];
-                        coords[3] = (float) current.y[1];
-                        coords[4] = (float) current.x[0];
-                        coords[5] = (float) current.y[0];
-                    }
+            numCoords = 2;
+            type = SEG_QUADTO;
+            coords[0] = previous.x[2];
+            coords[1] = previous.y[2];
+            coords[2] = current.x[0];
+            coords[3] = current.y[0];
+        } else {
+            numCoords = 3;
+            type = SEG_CUBICTO;
+            coords[0] = previous.x[2];
+            coords[1] = previous.y[2];
+            coords[2] = current.x[1];
+            coords[3] = current.y[1];
+            coords[4] = current.x[0];
+            coords[5] = current.y[0];
+        }
         this.numCoords = numCoords;
         this.type = type;
         this.doubleCoords = coords;
     }
-    
-    
-    
-    
-    
+
+  
+
     public int getNumCoords() {
         return numCoords;
     }
@@ -127,7 +126,7 @@ public class SegmentSeperator implements PathIterator {
     public void setType(int type) {
         this.type = type;
     }
-    
+
     public float[] getFloatCoords() {
         return floatCoords;
     }
@@ -168,5 +167,5 @@ public class SegmentSeperator implements PathIterator {
     public int currentSegment(double[] coords) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
-    
+
 }
